// src/shared/api/models/ClientRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientRead = {
    id: string;
    created: string;
    updated: string;
    slug: string;
    title: string;
    description?: (string | null);
    is_active?: boolean;
};